LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-to-srcdoc-expected.txt



FAIL navigate event fires appropriately (and can be canceled) for adding the srcdoc attribute assert_equals: expected (object) null but got (undefined) undefined
